refactor: switch form our `maps` package to `golang.org/x/exp`
Internal `maps` package was C/P of `golang.org/x/exp`
