+++ pkg/api/apis.go
@@ -17,7 +17,7 @@
 package api
 
 import (
-	"github.com/dynatrace/dynatrace-configuration-as-code/internal/maps"
+	"golang.org/x/exp/maps"
 )
 
 // APIs is a collection of API
