+++ pkg/manifest/manifest.go
@@ -16,7 +16,7 @@ package manifest
 
 import (
 	"fmt"
-	"github.com/dynatrace/dynatrace-configuration-as-code/internal/maps"
+	"golang.org/x/exp/maps"
 )
 
 type ProjectDefinition struct {
